+++ carbonmark-api/src/models/Methodology.model.ts
@@ -1,0 +1,7 @@
+import { Type } from "@sinclair/typebox";
+
+export const Methodology = Type.Object({
+  id: Type.String(),
+  category: Type.String(),
+  name: Type.String(),
+});
