+++ carbonmark-api/src/routes/projects/[id]/getById.schema.ts
@@ -1,0 +1,25 @@
+import { DetailedProject } from "../../../models/DetailedProject.model";
+
+export const schema = {
+  summary: "Project details",
+  description: "Retrieve a carbon project by its project ID",
+  tags: ["Projects"],
+  params: {
+    type: "object",
+    required: ["id"],
+    properties: {
+      id: {
+        type: "string",
+        description: "Project id & vintage",
+        examples: ["VCS-191-2008"],
+      },
+    },
+  },
+  response: {
+    "2xx": {
+      description: "Successful response",
+      type: "object",
+      properties: DetailedProject,
+    },
+  },
+};
