+++ carbonmark-api/src/routes/projects/get.ts
@@ -2,11 +2,11 @@ import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 import { mapValues, omit, sortBy } from "lodash";
 import { split } from "lodash/fp";
 import { FindProjectsQueryVariables } from "../../.generated/types/marketplace.types";
+import { ProjectT } from "../../models/Project.model";
 import { CreditId } from "../../utils/CreditId";
 import { gqlSdk } from "../../utils/gqlSdk";
 import { fetchAllCarbonProjects } from "../../utils/helpers/carbonProjects.utils";
 import { fetchAllPoolPrices } from "../../utils/helpers/fetchAllPoolPrices";
-import { CarbonmarkProjectT } from "./get.models";
 import { schema } from "./get.schema";
 import {
   CMSDataMap,
@@ -40,7 +40,7 @@ const handler = (fastify: FastifyInstance) =>
   async function (
     request: FastifyRequest<{ Querystring: Params }>,
     reply: FastifyReply
-  ): Promise<CarbonmarkProjectT[]> {
+  ): Promise<ProjectT[]> {
     //Transform the list params (category, country etc) provided so as to be an array of strings
     const args = mapValues(omit(request.query, "search"), split(","));
     //Get the default args to return all results unless specified
